refactor: change id type to long and database currency values - to numeric

+++ accounts-service/src/main/java/ru/avalc/accountsservice/repo/AccountsRepository.java
@@ -11,5 +11,5 @@ import ru.avalc.accountsservice.model.Account;
 @Repository
 public interface AccountsRepository extends CrudRepository<Account, Long> {
 
-    Account findByCustomerId(int customerId);
+    Account findByCustomerId(long customerId);
 }
