+++ cards-service/src/main/java/ru/avalc/cardsservice/repo/CardsRepository.java
@@ -13,5 +13,5 @@ import java.util.List;
 @Repository
 public interface CardsRepository extends CrudRepository<Card, Long> {
 
-    List<Card> findByCustomerId(int customerId);
+    List<Card> findByCustomerId(long customerId);
 }
